+++ src/briefy/leica/models/job/workflows.py
@@ -4,7 +4,6 @@ from briefy.common.vocabularies.roles import LocalRolesChoices as LR
 from briefy.common.workflow import BriefyWorkflow
 from briefy.common.workflow import Permission
 from briefy.common.workflow import WorkflowState as WS
-from briefy.leica.utils import transitions
 
 import logging