+++ src/briefy/leica/views/dashboard/professional.py
@@ -1,5 +1,4 @@
 """Views to handle Professional Dashboards."""
-from briefy.leica.models import Assignment
 from briefy.leica.models.dashboard.professional import DashboardProfessionalAssignment
 from briefy.ws import CORS_POLICY
 from briefy.ws.resources import RESTService
@@ -68,15 +67,12 @@ class DashboardProfessionalAssignmentService(RESTService):
         }
     )
 
-    @property
-    def default_filters(self) -> tuple:
+    def default_filters(self, query) -> object:
         """Default filters to be applied to every query.
 
         This is supposed to be specialized by resource classes.
         :returns: A tuple of default filters to be applied to queries.
         """
         user = self.request.user
-        filters = (
-            Assignment.professional_id == user.id,
-        )
-        return filters
+        query = query.params(professional_id_1=user.id)
+        return query